jsonrpc: make out_overflow NULL unless we need it.
Profiling showed we spent much of our time in tal_count(), so only
allocate it when we need it (which is rare):

Before:
	real	0m0.386000-0.415000(0.3936+/-0.0094)s

After:
	real	0m0.337000-0.370000(0.346+/-0.012)s

FWIW, here's the result With -O2:
	real	0m0.209000-0.291000(0.2299+/-0.021)s

At this point, we're spending 11% of our time in backend_wake, which
would require optimizations to ccan/io.

However, 64% of gossipd's CPU time is towire_gossip_getchannels_reply;
if we were less naive in our marshalling (we tal_resize()
continuously) we could cut that significantly, but I think we're fast
enough for now.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>