opening: Delay adding change outputs when funding fails
This was proposed in #2274, as a way to avoid chaining multiple
unconfirmed spends, by delaying the addition of change when we fail
the broadcast. The main difference is when we extract our change
output from a funding transaction and when we mark the reserved coins
a spent.

In case of failure, we'll still add the change script to the txfilter,
but wait for confirmation before adding the change output
itself. We'll also not mark the outputs as spent, but rather keep them
reserved.

Signed-off-by: Christian Decker <decker.christian@gmail.com>